Register :msgpack as a renderable content type
Enable a Rails server to respond to 'application/msgpack' requests with
a renderer that calls `to_msgpack`.

This commit has no tests and may not handle render options in an optimal
way, so it may be desirable to view it as a working proof-of-concept. It
does appear to work, though.